Refactor meetupa, obsługa wydarzeń online.
